feat/create preview component, clean upload component #70
